requirements.txt
streamlit>=1.22
pdfplumber>=0.7
python-docx>=0.8.11
pandas>=2.0
openai>=0.27.0
tqdm>=4.65
fuzzywuzzy[speedup]>=0.19.0
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"
numpy==1.26.4

